+++ app/src/main/java/com/example/sw_02/k_rst.java
@@ -3,7 +3,7 @@ package com.example.sw_02;
 public class k_rst {
     private String name;
     private String dong;
-    private int star;
+    private Double star;
 
     public k_rst(){}
 
@@ -23,11 +23,11 @@ public class k_rst {
         this.dong = dong;
     }
 
-    public int getStar() {
+    public Double getStar() {
         return star;
     }
 
-    public void setStar(int star) {
+    public void setStar(Double star) {
         this.star = star;
     }
 }
