+++ app/src/main/java/com/example/sw_02/s_rst.java
@@ -3,7 +3,7 @@ package com.example.sw_02;
 public class s_rst {
     private String dong;
     private String name;
-    private int star;
+    private Double star;
 
     public s_rst() {
     }
@@ -24,11 +24,11 @@ public class s_rst {
         this.name = name;
     }
 
-    public int getStar() {
+    public Double getStar() {
         return star;
     }
 
-    public void setStar(int star) {
+    public void setStar(Double star) {
         this.star = star;
     }
 }
